requirements.txt
# Web framework e servidor ASGI
fastapi
uvicorn[standard]

# ORM e driver assíncrono para MySQL
SQLAlchemy>=2.0
asyncmy

# Autenticação baseada em token (JWT)
python-jose[cryptography]

# Hash de senha seguro (recomendado): Argon2id via passlib
passlib
argon2-cffi

# Configuração por variáveis de ambiente (.env)
pydantic-settings
python-dotenv

# Migrações de banco (recomendado)
alembic
